fix: harden ctype predicates against invalid inputs and exclude surrogates
* `is_legalcharacter` now rejects surrogates (U+D800–U+DFFF), which are encoding artifacts, not Unicode scalar values
* `is_noncharacter` and the other `is_*` predicates now return 0 for negative or > U+10FFFF inputs instead of wrapping into the BMP or misreporting
* `totitlecase` now maps the full 32-bit code point range including supplementary planes (previously BMP-only)

Regenerated the Unicode 12.1 data tables with surrogates removed from the `islegal` bitmaps, added a libFuzzer harness (`fuzz/`) with `fuzz-ctype` and `fuzz-conversion` targets, and rewrote the ctype/conversion API docs.

+++ src/mulle-unicode-is-control.c
@@ -50,7 +50,7 @@ int   mulle_unicode16_is_control( uint16_t c)
 
 int   mulle_unicode_is_control( int32_t c)
 {
-   if( c <= 0xFFFF)
+   if( (uint32_t) c <= 0xFFFF)
       return( mulle_unicode16_is_control( (uint16_t) c));
 
    switch( c)
