+++ src/mulle-unicode-is-whitespaceornewline.c
@@ -56,7 +56,7 @@ int   mulle_unicode16_is_whitespaceornewline( uint16_t c)
 
 int   mulle_unicode_is_whitespaceornewline( int32_t c)
 {
-   if( c <= 0xFFFF)
+   if( (uint32_t) c <= 0xFFFF)
       return( mulle_unicode16_is_whitespaceornewline( (uint16_t) c));
    return( 0);
 }
